+++ src/basicradiocontroller.h
@@ -19,7 +19,7 @@ public:
   void stop() override;
   void pause() override;
 
-  void setVolume(qreal volume) override;
+  void setVolume(float volume) override;
   void setSource(const QUrl &source) override;
 
 private:
